+++ packages/typespec-rust/src/tcgcadapter/adapter.ts
@@ -7,7 +7,7 @@
 
 import * as codegen from '@azure-tools/codegen';
 import {values} from '@azure-tools/linq';
-import {DateTimeKnownEncoding, DiagnosticTarget, EmitContext, NoTarget} from '@typespec/compiler';
+import {DiagnosticTarget, EmitContext, NoTarget} from '@typespec/compiler';
 import * as http from '@typespec/http';
 import * as helpers from './helpers.js';
 import * as naming from './naming.js';
@@ -213,7 +213,7 @@ export class Adapter {
         throw new AdapterError('UnsupportedTsp', `unsupported enum underlying type ${sdkEnum.valueType.kind}`, sdkEnum.__raw?.node);
     }
 
-    rustEnum = new rust.Enum(enumName, sdkEnum.access === 'public', !sdkEnum.isFixed, enumType);
+    rustEnum = new rust.Enum(enumName, adaptAccessFlags(sdkEnum.access), !sdkEnum.isFixed, enumType);
     rustEnum.docs = this.adaptDocs(sdkEnum.summary, sdkEnum.doc);
     this.types.set(enumName, rustEnum);
 
@@ -445,7 +445,7 @@ export class Adapter {
     stack.push(unionName);
 
     const tspUnionProperties = this.getUnionProperties(union);
-    rustUnion = new rust.Union(unionName, union.access === 'public', tspUnionProperties.discriminatorPropertyName, tspUnionProperties.envelopePropertyName);
+    rustUnion = new rust.Union(unionName, adaptAccessFlags(union.access), tspUnionProperties.discriminatorPropertyName, tspUnionProperties.envelopePropertyName);
 
     rustUnion.docs = this.adaptDocs(union.summary, union.doc);
     this.types.set(unionName, rustUnion);
@@ -542,13 +542,18 @@ export class Adapter {
    * @returns the adapted Rust type
    */
   private getType(type: tcgc.SdkType, stack?: Array<string>): rust.Type {
-    const getDateTimeEncoding = (encoding: DateTimeKnownEncoding): rust.DateTimeEncoding => {
+    const getDateTimeEncoding = (encoding: string): rust.DateTimeEncoding => {
       switch (encoding) {
+        case 'rfc3339-fixed-width':
+          this.crate.addDependency(new rust.CrateDependency('time'));
+          return encoding;
         case 'rfc3339':
         case 'rfc7231':
           return encoding;
         case 'unixTimestamp':
           return 'unix_time';
+        default:
+          throw new AdapterError('UnsupportedTsp', `unhandled date-time encoding ${encoding}`, type.__raw?.node);
       }
     };
 
@@ -1316,7 +1321,7 @@ export class Adapter {
     methodOptionsStruct.fields.push(methodOptionsField);
 
 
-    const pub: rust.Visibility = method.access === 'public' ? 'pub' : 'pubCrate';
+    const pub: rust.Visibility = adaptAccessFlags(method.access);
     const methodOptions = new rust.MethodOptions(methodOptionsStruct);
     const httpMethod = method.operation.verb;
 
@@ -2419,6 +2424,15 @@ function hasClientNameDecorator(decorators: Array<tcgc.DecoratorInfo>): boolean 
  */
 function isHttpStatusCodeRange(statusCode: http.HttpStatusCodeRange | number): statusCode is http.HttpStatusCodeRange {
   return (<http.HttpStatusCodeRange>statusCode).start !== undefined;
+}
+
+/**
+ * converts tcgc's access flags (which aren't really flags) to visibility
+ * @param access the access flag to convert
+ * @returns the flag converted to visibility
+ */
+function adaptAccessFlags(access: tcgc.AccessFlags): rust.Visibility {
+  return access === 'public' ? 'pub' : 'pubCrate';
 }
 
 type QueryParamType = rust.QueryCollectionParameter | rust.QueryHashMapParameter | rust.QueryScalarParameter;